requirements.txt
flask==2.3.2
flask-babel==3.1.0
flask-sqlalchemy==3.0.3
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
spacy==3.7.2
python-docx==0.8.11
pdfminer.six==20221105